+++ src/cache.rs
@@ -1,23 +1,23 @@
-use crate::{
-    status::{get_status, Status},
-    Config, SentryApi,
-};
+use crate::status::Status;
 use async_trait::async_trait;
-use futures::future::{join_all, FutureExt};
-use primitives::{
-    BalancesMap, Channel, ChannelId,
-};
-use reqwest::Error;
-use slog::{error, info, Logger};
+use primitives::{BalancesMap, ChannelId};
+use slog::{info, Logger};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 use tokio::sync::RwLock;
-use url::Url;
+
+mod api_client;
+#[cfg(test)]
+pub mod mock_client;
+
+pub use api_client::ApiClient;
+#[cfg(test)]
+pub use mock_client::MockClient;
 
 // Re-export the Campaign
 pub use primitives::supermarket::Campaign;
 
-pub(crate)type Cached<T> = Arc<RwLock<T>>;
+pub(crate) type Cached<T> = Arc<RwLock<T>>;
 
 pub type ActiveCache = HashMap<ChannelId, Campaign>;
 pub type FinalizedCache = HashSet<ChannelId>;
@@ -29,77 +29,43 @@ pub enum ActiveAction {
     New(ActiveCache),
 }
 
-#[async_trait]
-pub trait CacheLike<'a>: core::fmt::Debug + Clone {
-    async fn initialize(logger: Logger, config: Config) -> Result<Self, Error>
-    where
-        Self: Sized;
-    async fn fetch_new_campaigns(&self);
-    async fn update(&self, new_active: ActiveAction, new_finalized: FinalizedCache);
-    async fn fetch_campaign_updates(&self);
-    async fn get_active_campaigns(&self) -> HashMap<ChannelId, Campaign>;
-}
+
+
+    #[async_trait]
+    pub trait Client: core::fmt::Debug + Clone {
+        fn logger(&self) -> Logger;
+        async fn collect_campaigns(&self) -> HashMap<ChannelId, Campaign>;
+        async fn fetch_campaign_updates(
+            &self,
+            active: &ActiveCache,
+        ) -> (ActiveAction, FinalizedCache);
+    }
 
 #[derive(Debug, Clone)]
-pub struct Cache {
+pub struct Cache<C: Client> {
     pub active: Cached<ActiveCache>,
     pub finalized: Cached<FinalizedCache>,
-    validators: HashSet<Url>,
+    client: C,
     logger: Logger,
-    sentry: SentryApi,
 }
 
-#[async_trait]
-impl<'a> CacheLike<'a> for Cache {
+impl<C> Cache<C> where C: Client {
     /// Fetches all the campaigns from the Validators on initialization.
-    async fn initialize(logger: Logger, config: Config) -> Result<Self, Error> {
-        info!(
-            &logger,
-            "Initialize Cache"; "validators" => format_args!("{:?}", &config.validators)
-        );
-        let sentry = SentryApi::new(config.timeouts.validator_request)?;
-
-        let validators = config.validators.clone();
+    pub async fn initialize(client: C) -> Self {
+        let logger = client.logger().clone();
+        info!(&logger, "Initialize Cache with Client"; "client" => ?&client);
 
         let cache = Self {
             active: Default::default(),
             finalized: Default::default(),
-            validators,
             logger,
-            sentry,
+            client,
         };
 
         // collect and initialize the active campaigns
         cache.fetch_new_campaigns().await;
 
-        Ok(cache)
-    }
-
-    /// # Update the Campaigns in the Cache with:
-    /// Collects all the campaigns from all the Validators and computes their Statuses
-    /// - New Campaigns
-    /// - New Finalized Campaigns
-    async fn fetch_new_campaigns(&self) {
-        let campaigns = collect_all_campaigns(&self.logger, &self.sentry, &self.validators).await;
-
-        let (active, finalized) = campaigns.into_iter().fold(
-            (HashMap::new(), HashSet::new()),
-            |(mut active, mut finalized), (id, campaign)| {
-                // we don't need to check if the ChannelId is already in either active or finalized,
-                // since the Campaigns (ChannelId) cannot repeat in the HashMap
-                match &campaign.status {
-                    Status::Finalized(_) => {
-                        finalized.insert(id);
-                    }
-                    _ => {
-                        active.insert(id, campaign);
-                    }
-                }
-                (active, finalized)
-            },
-        );
-
-        self.update(ActiveAction::New(active), finalized).await
+        cache
     }
 
     /// Updates the full Cache with the new values:
@@ -164,6 +130,33 @@ impl<'a> CacheLike<'a> for Cache {
         } // Finalized cache - release of RwLockWriteGuard
     }
 
+    /// # Update the Campaigns in the Cache with:
+    /// Collects all the campaigns from all the Validators and computes their Statuses
+    /// - New Campaigns
+    /// - New Finalized Campaigns
+    pub async fn fetch_new_campaigns(&self) {
+        let campaigns = self.client.collect_campaigns().await;
+
+        let (active, finalized) = campaigns.into_iter().fold(
+            (HashMap::new(), HashSet::new()),
+            |(mut active, mut finalized), (id, campaign)| {
+                // we don't need to check if the ChannelId is already in either active or finalized,
+                // since the Campaigns (ChannelId) cannot repeat in the HashMap
+                match &campaign.status {
+                    Status::Finalized(_) => {
+                        finalized.insert(id);
+                    }
+                    _ => {
+                        active.insert(id, campaign);
+                    }
+                }
+                (active, finalized)
+            },
+        );
+
+        self.update(ActiveAction::New(active), finalized).await
+    }
+
     /// Reads the active campaigns and schedules a list of non-finalized campaigns
     /// for update from the Validators
     ///
@@ -172,136 +165,28 @@ impl<'a> CacheLike<'a> for Cache {
     /// - Add to the Finalized cache
     /// Other statuses:
     /// - Update the Status & Balances from the latest Leader NewState
-    async fn fetch_campaign_updates(&self) {
-        // we need this scope to drop the Read Lock on `self.active`
-        // before performing the finalize & update actions
-        let (update, finalize) = {
-            let active = self.active.read().await;
+    pub async fn fetch_campaign_updates(&self) {
+        let (active_action, finalized) = self
+            .client
+            .fetch_campaign_updates(&*self.active.read().await)
+            .await;
 
-            let mut update = HashMap::new();
-            let mut finalize = HashSet::new();
-            for (id, campaign) in active.iter() {
-                match get_status(&self.sentry, &campaign.channel).await {
-                    Ok((Status::Finalized(_), _balances)) => {
-                        finalize.insert(*id);
-                    }
-                    Ok((new_status, new_balances)) => {
-                        update.insert(*id, (new_status, new_balances));
-                    }
-                    Err(err) => error!(
-                        &self.logger,
-                        "Error getting Campaign ({:?}) status", id; "error" => ?err
-                    ),
-                };
-            }
-
-            (update, finalize)
-        };
-
-        self.update(ActiveAction::Update(update), finalize).await;
+        self.update(active_action, finalized).await
     }
-
-    async fn get_active_campaigns(&self) -> HashMap<ChannelId, Campaign> {
-        // Needed to access the campaigns from RwLock
-        {
-            let read_campaigns = self.active.read().await;
-            let mut active = HashMap::new();
-            for (id, campaign) in read_campaigns.iter() {
-                active.insert(*id, campaign.clone());
-            }
-            active
-        }
-    }
-}
-
-async fn collect_all_campaigns(
-    logger: &Logger,
-    sentry: &SentryApi,
-    validators: &HashSet<Url>,
-) -> HashMap<ChannelId, Campaign> {
-    let mut campaigns = HashMap::new();
-
-    for channel in get_all_channels(logger, sentry, validators).await {
-        /*
-            @TODO: Issue #23 Check ChannelId and the received Channel hash
-            We need to figure out a way to distinguish between Channels, check if they are the same and to remove incorrect ones
-            For now just check if the channel is already inside the fetched channels and log if it is
-        */
-        // if campaigns.contains_key(&channel.id) {
-        //     info!(
-        //         logger,
-        //         "Skipping Campaign ({:?}) because it's already fetched from another Validator",
-        //         &channel.id
-        //     )
-        // } else {
-        match get_status(&sentry, &channel).await {
-            Ok((status, balances)) => {
-                let channel_id = channel.id;
-                campaigns
-                    .entry(channel_id)
-                    .and_modify(|campaign: &mut Campaign| {
-                        campaign.status = status.clone();
-                        campaign.balances = balances.clone();
-                    })
-                    .or_insert_with(|| Campaign::new(channel, status, balances));
-            }
-            Err(err) => error!(
-                logger,
-                "Failed to fetch Campaign ({:?}) status from Validator", channel.id; "error" => ?err
-            ),
-        }
-        // }
-    }
-
-    campaigns
-}
-
-/// Retrieves all channels from all Validator URLs
-async fn get_all_channels<'a>(
-    logger: &Logger,
-    sentry: &SentryApi,
-    validators: &HashSet<Url>,
-) -> Vec<Channel> {
-    let futures = validators.iter().map(|validator| {
-        sentry
-            .get_validator_channels(validator)
-            .map(move |result| (validator, result))
-    });
-
-    join_all(futures)
-    .await
-    .into_iter()
-    .filter_map(|(validator, result)| match result {
-            Ok(channels) => {
-                info!(logger, "Fetched {} active Channels from Validator ({})", channels.len(), validator);
-
-                Some(channels)
-            },
-            Err(err) => {
-                error!(logger, "Failed to fetch Channels from Validator ({})", validator; "error" => ?err);
-
-                None
-            }
-        })
-    .flatten()
-    .collect()
 }
 
 #[cfg(test)]
 mod test {
     use super::*;
 
-    use crate::status::test::{get_approve_state_msg, get_heartbeat_msg, get_new_state_msg};
+    use crate::{SentryApi, status::test::{get_approve_state_msg, get_heartbeat_msg, get_new_state_msg}};
     use crate::{config::DEVELOPMENT, util::test::discard_logger};
     use chrono::{Duration, Utc};
-    use primitives::{
-        sentry::{
+    use primitives::{Channel, sentry::{
             ChannelListResponse, LastApproved, LastApprovedResponse, NewStateValidatorMessage,
             ValidatorMessage, ValidatorMessageResponse,
-        },
-        util::tests::prep_db::{DUMMY_CHANNEL, DUMMY_VALIDATOR_FOLLOWER, DUMMY_VALIDATOR_LEADER},
-        validator::{MessageTypes, NewState},
-    };
+        }, util::tests::prep_db::{DUMMY_CHANNEL, DUMMY_VALIDATOR_FOLLOWER, DUMMY_VALIDATOR_LEADER}, validator::{MessageTypes, NewState}};
+        use reqwest::Url;
     use wiremock::{
         matchers::{method, path, query_param},
         Mock, MockServer, ResponseTemplate,
@@ -311,19 +196,22 @@ mod test {
         active: HashMap<ChannelId, Campaign>,
         finalized: HashSet<ChannelId>,
         validators: HashSet<Url>,
-    ) -> Result<Cache, Box<dyn std::error::Error>> {
+    ) -> Result<Cache<ApiClient>, Box<dyn std::error::Error>> {
         let sentry = SentryApi::new(std::time::Duration::from_secs(60))?;
         let logger = discard_logger();
 
-        let cache = Cache {
-            active: Arc::new(RwLock::new(active)),
-            finalized: Arc::new(RwLock::new(finalized)),
+        let client = ApiClient {
             logger,
             sentry,
             validators,
         };
 
-        Ok(cache)
+        Ok(Cache {
+            active: Arc::new(RwLock::new(active)),
+            finalized: Arc::new(RwLock::new(finalized)),
+            logger: client.logger().clone(),
+            client,
+        })
     }
 
     fn setup_channel(leader_url: &Url, follower_url: &Url) -> Channel {
@@ -449,9 +337,11 @@ mod test {
             .mount(&mock_server)
             .await;
 
-        let cache = Cache::initialize(discard_logger(), config)
+
+        let client = ApiClient::init(discard_logger(), config)
             .await
             .expect("Should initialize");
+        let cache = Cache::initialize(client).await;
 
         let active_cache = cache.active.read().await;
         let active_campaign = active_cache
